New Functions for google app scripts (#49)
* New Functions for google app scripts

Added functions to get google app project, and send request body to functions in .gs scripts as part of those projects.

* Tweak invoke

Added -useBasicParsing to invoke-restmethod.

Added parameter sets. One to use a psobject/hashtable and a second to just provide the options for the request body.

Added script validation on $requestBody

* Modify DevMode

Modify devmode to be string instead of bolean. Validate option set, and provided an additional example.

* DevMode switch