Update UMN-Google.psm1
The contenttype for Invoke-RestMethod in Set-GSheetData was set to "application/json". That was fine as long as you didn't use umlauts like ä or é. 
I added a parameter that allows to change the contenttype for the request so you can transmit umlauts with the contenttype 'application/json; charset=utf-8' with the parameter -contenttype 'application/json; charset=utf-8'.

The default is still 'application/json' for backwards compability.